Extract Terraform command registration
This extracts registering Terraform commands to a central class. This cleans up extension.ts from having lots of registrations, enables a central place to keep track of terraform-ls execution related commands, and enables centralized progress reporting in a future commit.
